requirements.txt
numpy==1.26.2
opencv-python==4.8.1.78
pyserial==3.5
